+++ include/asm-generic/types.h
@@ -8,7 +8,7 @@
 
 #ifndef __ASSEMBLY__
 
-typedef unsigned short umode_t;
+typedef unsigned int umode_t;
 
 #endif /* __ASSEMBLY__ */
